% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDEGsFunction.r
\name{calculateDEGs}
\alias{calculateDEGs}
\title{This script imports & filters the data and exports it to a .csv file
Should we add a prompt for the user to specify the files used?
This documentation needs updating later}
\usage{
calculateDEGs(input_file, output_file, sample_groups)
}
\description{
This script imports & filters the data and exports it to a .csv file
Should we add a prompt for the user to specify the files used?
This documentation needs updating later
}
