% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichKEGGING.R
\name{enrichKEGGING}
\alias{enrichKEGGING}
\title{enrichKEGG}
\usage{
enrichKEGGING(input_file_1, input_file_2, output_file)
}
\arguments{
\item{input_file_1}{This is the path to the file containing the differentially expressed genes (DEGs) for the genes input in enrichKEGG.
The data in this file should be pre-processed to filter out low-expressed genes and adjust for log fold change.}

\item{input_file_2}{This is the path to the file containing the differentially expressed genes (DEGs) for the universe input in enrichKEGG.
The data in this file should be pre-processed to filter out low-expressed genes, adjust for log fold change, p-value, and FDR.}

\item{output_file}{This is the path to the .xlsx file that will be created, containing the results of enrichKEGG analysis..}
}
\description{
Prepares data obtained from DEG calculation to be suitable for Gene Enrichment Analysis using the enrichKEGG function of the clusterprofiler package.
 The two files are post filtered and log fold changed DEG data file which is input in the Universe parameter and fully DEG processed file including filtration, log fold change, padjustment and FDR correction.
 The latter file is used in the 'genes' attribute of enrichKEGG. The enrichKEGG object is used to generate the barplot, dotplot and upset plot for enrichment visualisation.
}
\examples{
library(DEGpathwayAnalysis)





}
